% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_bullet.R
\name{plot_bullet}
\alias{plot_bullet}
\title{plot_bullet()
#'  Custom version of the classical Bullet Graph
 \url{http://en.wikipedia.org/wiki/Bullet_graph}}
\usage{
plot_bullet(df, my_title = "Overview of JRT results for FY14\\n",
  my_palette = c("#B10318", "#DBA13A", "#309343"), my_font = "Impact",
  my_background = "#E2E2E3")
}
\arguments{
\item{df:}{data.frame with 1 rows and the following 7 columns
measure  : label of what's being measured
target   : the target value for the measure
value    : the actual value of the measure
n        : Sample size}
}
\value{
ggplot object
}
\description{
plot_bullet()
#'  Custom version of the classical Bullet Graph
 \url{http://en.wikipedia.org/wiki/Bullet_graph}
}
\examples{
df <- data.frame(measure = c("\% enrolled", "\% assessed", "\% gaining skills"),
target=c(.75, .9, .72),
value=c(232, 50, 36),
n = c(407, 232, 50))

plot_bullet(df)
}

