% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_singbullet.R
\name{plot_singbullet}
\alias{plot_singbullet}
\title{plot_singbullet()
#'  Custom version of the classical Bullet Graph
 \url{http://en.wikipedia.org/wiki/Bullet_graph}
 This function plots only the body of a single bullet chart: No text}
\usage{
plot_singbullet(df, my_palette = c(low = "#B10318", medium = "#DBA13A", high =
  "#309343"), my_font = "Impact", my_background = "#E2E2E3")
}
\arguments{
\item{df:}{data.frame with 1 rows and the following 7 columns
measure  : label of what's being measured
target   : the target value for the measure
value    : the actual value of the measure
n        : Sample size}
}
\value{
ggplot object
}
\description{
plot_singbullet()
#'  Custom version of the classical Bullet Graph
 \url{http://en.wikipedia.org/wiki/Bullet_graph}
 This function plots only the body of a single bullet chart: No text
}
\examples{
df <- data.frame(measure = c("\% enrolled", "\% assessed", "\% gaining skills"),
target=c(.75, .9, .72),
value=c(232, 50, 36),
n = c(407, 232, 50))

plot_singbullet(df[1, ])
}

